+++ src/extensions/plugin/loaders.ts
